mockup exceptions and Ids changed from Integer to Long

+++ wdm-microservices-project/orders/src/main/java/wdm/project/dto/OrderItem.java
@@ -22,7 +22,7 @@ public class OrderItem implements Serializable {
         return id;
     }
 
-    public void setId(Integer orderId, Integer itemId) {
+    public void setId(Long orderId, Long itemId) {
         this.id.setOrderId(orderId);
         this.id.setItemId(itemId);
     }
