+++ wdm-microservices-project/orders/src/main/java/wdm/project/repository/OrdersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import wdm.project.dto.Order;
 
 @Repository
-public interface OrdersRepository extends JpaRepository<Order, Integer> {
+public interface OrdersRepository extends JpaRepository<Order, Long> {
 }
